+++ client/src/view/customElements/index.ts
@@ -5,6 +5,8 @@ import AppHeader from './app-header';
 import CalendarCell from './calendar-cell';
 import CategorySelect from './category-select';
 import MethodSelect from './method-select';
+import HistoryForm from './history-form';
+import HistoryList from './history-list';
 
 RouterLink.define();
 ListTitle.define();
@@ -13,3 +15,5 @@ AppHeader.define();
 CalendarCell.define();
 CategorySelect.define();
 MethodSelect.define();
+HistoryForm.define();
+HistoryList.define();
